retailseller/retailseller/retailseller/customer/requirements.txt
django
ckeditor
django-phonenumberfield
